New client file
Split cerver file into cerver and client files. Also refactor some code with my typedefs
