+++ include/cerver.h
@@ -5,45 +5,21 @@
     #define _GNU_SOURCE
 #endif
 
-#include <stdio.h>
-#include <stdint.h>
+#include <poll.h>
+
+#include "myTypes.h"
 
 #include "network.h"
-#include "game.h"       // game server
+#include "client.h"
+#include "game.h"
 
 #include "utils/list.h"
-#include "utils/config.h"
 #include "utils/objectPool.h"
-// #include "utils/vector.h"
+#include "utils/avl.h"
 
-#include <poll.h>
-#include "utils/avl.h"      // 02/11/2018 -- using an avl tree to handle clients
+#include "utils/config.h"
 
-// 04/11/2018 -- 23:04 -- including a th pool inside each server
 #include "utils/thpool.h"
-
-#define EXIT_FAILURE    1
-
-#define THREAD_OK       0
-
-typedef uint8_t u8;
-typedef uint16_t u16;
-typedef uint32_t u32;
-typedef uint64_t u64;
-
-typedef int8_t i8;
-typedef int16_t i16;
-typedef int32_t i32;
-typedef int64_t i64;
-
-typedef unsigned char asciiChar;
-
-// takes no argument and returns a value (int)
-typedef u8 (*Func)(void);
-// takes an argument and does not return a value
-typedef void (*Action)(void *);
-// takes an argument and returns a value (int)
-typedef u8 (*delegate)(void *);
 
 #define DEFAULT_USE_IPV6                0
 #define DEFAULT_PROTOCOL                IPPROTO_TCP
@@ -63,38 +39,6 @@ typedef u8 (*delegate)(void *);
 #define poll_n_fds      100           // n of fds for the pollfd array
 
 struct _Server;
-
-/*** CLIENT ***/
-
-#pragma region CLIENT
-
-// anyone that connects to the server
-struct _Client {
-
-    // 25/11/2018 - 16:00 - using connection values as the client id
-    char *clientID;
-    struct sockaddr_storage address;
-
-    u8 authTries;           // remaining attemps to authenticate
-    bool dropClient;        // client failed to authenticate
-
-    // 24/11/2018 -- thinking on a kind of web sessions
-    char *sessionID;            // generated by the server for each client
-    // a client may need to have multiple connections at the same time...
-    u8 n_active_cons;
-    i32 *active_connections;   
-
-};
-
-typedef struct _Client Client;
-
-extern Client *getClientBySocket (AVLNode *node, i32 socket_fd);
-extern Client *getClientBySession (AVLTree *clients, char *sessionID);
-
-extern void client_registerToServer (struct _Server *server, Client *client, int);
-extern void client_unregisterFromServer (struct _Server *server, Client *client);
-
-#pragma endregion
 
 /*** SEVER ***/
